+++ CHANGELOG.rst
@@ -54,7 +54,8 @@ Added
   keyboard shortcut. Optionally, you can also have the next untagged image selected automatically
   afterwards. Thanks to Robert Menger for the respective feature request :-)
 
-* Bookmarks can now be added using the map's context menu.
+* Added more functionality to the map's context menu: Bookmarks can now be added for the current
+  map center and it's coordinates can be copied to the clipboard.
 
 Changed
 =======